backend/requirements.txt
FLASK==2.2.3
python-dateutil==2.8.2
requests==2.28.2